requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
protobuf==4.21.10
SQLAlchemy==1.4.44
Werkzeug==2.0.0
WTForms==3.0.1
